requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.1.0
bs4==0.0.2
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.10
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
mysql-connector-python==9.1.0
numpy==2.0.2
packaging==24.2
praw==7.8.1
prawcore==2.4.0
preshed==3.0.9
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.1.0
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.5.0
thinc==8.3.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
